гет лабворкс + гет лабворк бай айди + обработка ошибок

+++ src/main/java/com/vezhur/soa/repository/LabworkRepository.java
@@ -5,6 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LabworkRepository extends JpaRepository<LabworkEntity, Long> {
-
-}
+public interface LabworkRepository extends JpaRepository<LabworkEntity, Integer> {}
